Data sample ported to 64-bits

+++ Data/Screens/ADONET/BasicOperations.cs
@@ -2,8 +2,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using System.Data;
 using System.IO;
 using Mono.Data.Sqlite;
@@ -138,9 +138,8 @@ namespace Xamarin.Screens.ADONET
 			
 			public TableSource(List<string> items) : base() { this.items = items; }
 			
-			public override int NumberOfSections (UITableView tableView) { return 1; }
-			
-			public override int RowsInSection (UITableView tableview, int section) { return this.items.Count; }
+			public override nint NumberOfSections (UITableView tableView) { return 1; }
+			public override nint RowsInSection (UITableView tableview, nint section) { return this.items.Count; }
 			
 			public override UITableViewCell GetCell (UITableView tableView, NSIndexPath indexPath)
 			{
@@ -148,7 +147,7 @@ namespace Xamarin.Screens.ADONET
 				cell = tableView.DequeueReusableCell("item");
 				if(cell == null) 
 					cell = new UITableViewCell(UITableViewCellStyle.Default, "item");
-				cell.TextLabel.Text = this.items[indexPath.Row];
+				cell.TextLabel.Text = this.items[(int)indexPath.Row];
 				return cell;
 			}
