+++ Data/Screens/ViciCoolStorage/BasicOperations.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using System.Data;
 using System.IO;
 using Vici.CoolStorage;
@@ -108,9 +108,9 @@ namespace Xamarin.Screens.ViciCoolStorage
 			
 			public TableSource (CSList<Person> items) : base() { this.items = items; }
 			
-			public override int NumberOfSections (UITableView tableView) { return 1; }
+			public override nint NumberOfSections (UITableView tableView) { return 1; }
 			
-			public override int RowsInSection (UITableView tableview, int section) { return this.items.Count; }
+			public override nint RowsInSection (UITableView tableview, nint section) { return this.items.Count; }
 			
 			public override UITableViewCell GetCell (UITableView tableView, NSIndexPath indexPath)
 			{
@@ -118,7 +118,7 @@ namespace Xamarin.Screens.ViciCoolStorage
 				cell = tableView.DequeueReusableCell ("item");
 				if (cell == null)
 					cell = new UITableViewCell (UITableViewCellStyle.Default, "item");
-				cell.TextLabel.Text = this.items[indexPath.Row].FirstName + " " + this.items[indexPath.Row].LastName;
+				cell.TextLabel.Text = this.items[(int)indexPath.Row].FirstName + " " + this.items[(int)indexPath.Row].LastName;
 				return cell;
 			}
